(1248) Split up organizationModel
