+++ packages/api/server/models/organizationModel/findAssociationName.js
@@ -1,0 +1,23 @@
+const { sequelize } = require('#db/models');
+
+module.exports = async (name) => {
+    const result = await sequelize.query(
+        `SELECT
+            organizations.name,
+            organizations.abbreviation
+        FROM localized_organizations AS organizations
+        LEFT JOIN
+            organization_types ON organizations.fk_type = organization_types.organization_type_id
+        WHERE
+            organization_types.fk_category = 'association'
+            AND
+            organizations.name ILIKE :name`,
+        {
+            type: sequelize.QueryTypes.SELECT,
+            replacements: {
+                name,
+            },
+        },
+    );
+    return result.length > 0 ? result[0] : null;
+};
