+++ packages/api/server/models/organizationModel/findOneById.js
@@ -1,0 +1,22 @@
+const { sequelize } = require('#db/models');
+
+module.exports = async (id) => {
+    const result = await sequelize.query(
+        `SELECT
+            organizations.organization_id AS id,
+            organizations.name,
+            organizations.fk_type,
+            organization_types.fk_category
+        FROM organizations
+        LEFT JOIN organization_types ON organizations.fk_type = organization_types.organization_type_id
+        WHERE organizations.organization_id = :id`,
+        {
+            type: sequelize.QueryTypes.SELECT,
+            replacements: {
+                id,
+            },
+        },
+    );
+
+    return result.length === 1 ? result[0] : null;
+};
